sketches DataParser and cocoparser
