+++ mantisshrimp/parsers/coco_parsers/__init__.py
@@ -1,1 +1,2 @@
 from .coco_image_info_parser import *
+from .coco_annotation_parser import *
